Add specific configuration for timidity and freepat
Ships freepat as part of packages included.
Include a timidity config which is necessary for timidity to start inside the
snap. Before, it was relying on /etc/timidity/timidity.cfg which was actually
accessible in the snap if timidity was installed on the host.
(Referenced as https://bugs.launchpad.net/snappy/+bug/1648712)